전남대 BE_유보민 6주차 과제 (2단계) (#268)
* Move from spring-gift-order

* refactor : 필드 수정

* refactor : 필드 수정

* refactor : 대체키 설정, 필드 수정

* delete : 복합키 삭제

* refactor : 카테고리 생성, 수정 추가

* refactor : 카테고리 생성, 수정 추가

* refactor : 생성자 인자 추가

* refactor : import 삭제

* refactor : 필드 삭제로 인한 put 삭제

* refactor : 상품 수량 삭제로 변경

* refactor : 로그인 폼 수정

* refactor : 필드 수정으로 인해 생성자에 인자 추가

* refactor : 필드 수정으로 인해 생성자에 인자 추가, 수정

* refactor : 반환 타입 수정

* refactor : 오버라이딩한 메서드 삭제, 수정

* feat : 주문 Controller 추가

* refactor : 메서드명 수정, 기능 수정

* docs : 1단계 README.md 작성

* feat : 배포 스크립트 작성

* docs : 2단계 요구사항 추가

* fix : BUILD_PATH 수정

* refactor : HiddenHttpMethodFilter 추가

* refactor : 추가/수정에 대한 데이터 반환으로 수정

* refactor : CategoryDto 반환으로 수정

* refactor : categoryService 메서드의 반환 타입 수정으로 인해 함께 수정

* refactor : redirect 삭제

* refactor : OptionService의 반환 타입 수정, 영향 받는 메서드의 변수 타입들 수정

* refactor : 금액 계산 로직을 OrderService에서 수행하도록 수정

* refactor : 상품 추가/수정 후 데이터를 반환하도록 수정, Service의 메서드들도 Dto 반환하도록 수정

* refactor : OrderService 추가, 금액 계산 로직 옮겨오기

* refactor : 로그 보기

+++ src/main/java/gift/Controller/ProductController.java
@@ -16,8 +16,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -55,28 +55,26 @@ public class ProductController {
     @PostMapping("/")
     @Operation(summary = "상품 생성", description = "상품을 생성합니다.")
     public ResponseEntity<?> createProduct(@Valid @RequestBody ProductDto productDto) {
-        productService.saveProduct(productDto);
-        URI location = ServletUriComponentsBuilder.fromCurrentRequest()
-                .path("/")
-                .buildAndExpand(productDto.getId())
-                .toUri();
-        return ResponseEntity.created(location).body("Product added successfully");
+        ProductDto savedproductDto = productService.saveProduct(productDto);
+        return ResponseEntity.status(HttpStatus.CREATED).body(savedproductDto);
     }
 
     @PutMapping("/{productId}")
     @Operation(summary = "상품 수정", description = "상품에 대한 정보를 수정합니다.")
-    public ResponseEntity<?> updateProductById(@PathVariable Long productId, @Valid @ModelAttribute ProductDto productDtoDetails) {
-        Product product = productService.updateProduct(productDtoDetails);
-        return ResponseEntity.ok(product);
+    public ResponseEntity<?> updateProductById(@PathVariable Long productId, @Valid @RequestBody ProductDto productDtoDetails) {
+        ProductDto productDto = productService.updateProduct(productDtoDetails);
+        return ResponseEntity.ok(productDto);
     }
 
     @DeleteMapping("/{productId}")
     @Operation(summary = "상품 삭제", description = "상품을 삭제합니다.")
-    public ResponseEntity<?> deleteProduct(@PathVariable Long productId, Model model) {
+    public ResponseEntity<?> deleteProduct(@PathVariable Long productId) {
         Optional<ProductDto> optionalProduct = productService.getProductById(productId);
-        model.addAttribute("product", optionalProduct.get());
+        if (optionalProduct.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Product not found");
+        }
         productService.deleteProduct(productId);
-        return ResponseEntity.ok(null);
+        return ResponseEntity.ok("Product deleted successfully");
     }
 
 }
